add get_iceberg_ddl_options back to be backwards compatible with original iceberg table configuration
