+++ dbt-snowflake/src/dbt/adapters/snowflake/relation.py
@@ -1,3 +1,5 @@
+import textwrap
+
 from dataclasses import dataclass, field
 from typing import FrozenSet, Optional, Type, Iterator, Tuple
 
@@ -200,6 +202,23 @@ class SnowflakeRelation(BaseRelation):
             return "iceberg"
         else:
             return ""
+
+    def get_iceberg_ddl_options(self, config: RelationConfig) -> str:
+        # If the base_location_root config is supplied, overwrite the default value ("_dbt/")
+        base_location: str = (
+            f"{config.get('base_location_root', '_dbt')}/{self.schema}/{self.name}"  # type:ignore
+        )
+
+        if subpath := config.get("base_location_subpath"):  # type:ignore
+            base_location += f"/{subpath}"
+
+        external_volume = config.get("external_volume")  # type:ignore
+        iceberg_ddl_predicates: str = f"""
+        external_volume = '{external_volume}'
+        catalog = 'snowflake'
+        base_location = '{base_location}'
+        """
+        return textwrap.indent(textwrap.dedent(iceberg_ddl_predicates), " " * 10)
 
     def __drop_conditions(self, old_relation: "SnowflakeRelation") -> Iterator[Tuple[bool, str]]:
         drop_view_message: str = (
